// apps/chat-service/src/group-chat/group-chat.enum.ts
export enum EGroupChatRoles {
  ADMIN = 'ADMIN',
  MEMBER = 'MEMBER',
}

export enum EGroupChatPermissions {
  SEND_MESSAGE = 'sendMessage',
  PIN_MESSAGE = 'pinMessage',
  SHARE_INVITE_CODE = 'shareInviteCode',
  UPDATE_INFO = 'updateInfo',
}

export enum EJoinRequestStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
}